Adding and fixing training - incomplete
It seems like there's an issue in actually calculating the derivative. The loss decreases, but at a much lower speed than it should be; in addition. it seems like some sections of biases or weights are simply not responding. This suggests problems to be found in the derivative functions of the Layer class, which will be promptly investigated and resolved.